stm32 usb: fix disconnect-gpios behaviour.
For some reason, rather than testing for the presence of
disconnect-gpios in the devicetree, the USB Device driver was relying
on a hidden Kconfig flag to be set.

This patch removes the Kconfig option completely and simply tests for
the DT property - if it's set, you obviously know what you're doing and
obviously need the pull-up GPIO behaviour.